Secure the potential NPE

+++ main/plugins/org.talend.utils/src/main/java/org/talend/utils/files/FileUtils.java
@@ -188,18 +188,20 @@ public final class FileUtils {
      * @param func a {@link Function} that will be used to filter on the files to delete, according to their name
      */
     public static void deleteFiles(File folder, Function<String, Boolean> func) {
-        if (folder.exists()) {
-            FilenameFilter filter = new FilenameFilter() {
+        if (folder != null) {
+            if (folder.exists()) {
+                FilenameFilter filter = new FilenameFilter() {
 
-                @Override
-                public boolean accept(File _dir, String name) {
-                    return func.apply(name);
+                    @Override
+                    public boolean accept(File _dir, String name) {
+                        return func != null && func.apply(name);
+                    }
+
+                };
+                List<File> filesToRemove = getAllFilesFromFolder(folder, filter);
+                for (File fileToRemove : filesToRemove) {
+                    fileToRemove.delete();
                 }
-
-            };
-            List<File> filesToRemove = getAllFilesFromFolder(folder, filter);
-            for (File fileToRemove : filesToRemove) {
-                fileToRemove.delete();
             }
         }
     }
